basestructure (#4)
Base file structure for back-end.

-----

Repository / handler packages separated in such a way that it is possible to switch to different db as along as it has all the necessary repository functions defined in models(interfaces).

Code still contains exemple code snippets for handler, interface, repository and db initialization. Snippets are supposed to be deleted after real code is written, till then it stays there as example. Snippets are targeted with //TEST or //code example comment to be easly found.

Co-authored-by: CitaZane <zanekrumina2@gmail.com>
Reviewed-on: https://01.kood.tech/git/itskris/social-network/pulls/4
Co-authored-by: CitaZane <citazane@noreply.01.kood.tech>
Co-committed-by: CitaZane <citazane@noreply.01.kood.tech>