Mirations (#18)
### Purpose of changes
Create database migrations. Find the plan for the database at [This link](https://drawsql.app/kood-j-hvi-2/diagrams/social-network#).

### What are migrations
Migrations are like git commits for databases. The purpose of a up migration is to change the structure of a database by adding, changing or removing tabels. If after applying some changes the database broke, we can remove those changes with a down migration and try again. That way there's less data loss.

### How it works
in sqlite.go, after successfully finding the database, the migrations are called on that database. in sqlire > Migration we locate the folder with all the up migrations and we execute them on the database if necessary (it's similar to if we started with a plank project and then apply all the git commits one by one untill we had a finished project). We return an error if we couldn't find the migrations folder of if there was an error during the migration process.

Co-authored-by: Laura-Eliise <laura-eliise@marrandi.ee>
Reviewed-on: https://01.kood.tech/git/itskris/social-network/pulls/18